Reports has a new folder and file.
