+++ source/constants/blueprints.py
@@ -4,7 +4,7 @@ from typing import Final
 INDEX_BLUEPRINT_NAME: Final[str] = "index"
 GAME_STATS_BLUEPRINT_NAME: Final[str] = "game_stats"
 PROFILE_BLUEPRINT_NAME: Final[str] = "profile"
-REPORTS_TABLE_NAME: Final[str] = "report"
+REPORT_BLUEPRINT_NAME: Final[str] = "report"
 USER_BLUEPRINT_NAME: Final[str] = "user"
 XP_EVENT_TYPE_BLUEPRINT_NAME: Final[str] = "xp_event_type"
 AUTH_BLUEPRINT_NAME: Final[str] = "auth"
